chatbot-hr/chatbot-hr-legal-l1-l2/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
langchain==0.0.326
azure-storage-blob==12.19.0
python-multipart==0.0.6  # For FastAPI file uploads
